added a rule engine. ANd the next TODOs

+++ src/main/java/com/example/ordersService/domain/Order.java
@@ -35,5 +35,5 @@ public class Order {
     @Column(name = "quantity")
     private Map<String, Integer> productCodeQuantityMap;
 
-    private Double priceInCents;
+    private Integer priceInCents;
 }
